+++ demisto_sdk/commands/format/update_mapper.py
@@ -5,7 +5,7 @@ import click
 from demisto_sdk.commands.common.constants import \
     LAYOUT_AND_MAPPER_BUILT_IN_FIELDS
 from demisto_sdk.commands.common.tools import \
-    get_all_incident_and_indicator_fields_from_id_set
+    get_all_incident_and_indicator_fields_from_id_set, get_invalid_incident_fields_from_mapper
 from demisto_sdk.commands.common.update_id_set import BUILT_IN_FIELDS
 from demisto_sdk.commands.format.format_constants import (ERROR_RETURN_CODE,
                                                           SKIP_RETURN_CODE,
@@ -39,7 +39,7 @@ class MapperJSONFormat(BaseUpdateJSON):
             self.set_description()
             self.set_mapping()
             self.update_id()
-            self.remove_inexistent_fields()
+            self.remove_non_existent_fields()
             self.save_json_to_destination_file()
             return SUCCESS_RETURN_CODE
 
@@ -62,40 +62,24 @@ class MapperJSONFormat(BaseUpdateJSON):
         if not self.data.get('mapping'):
             self.data['mapping'] = {}
 
-    def extract_content_fields(self, content_fields, built_in_fields):
+    def remove_non_existent_fields(self):
         """
-        Extract fields which only exist in the id set file.
+        Remove non-existent fields from a mapper.
         """
-        def _extract_content_fields(field):
-            inc_name, inc_info = field
-            # incoming mapper
-            if self.data.get('type', {}) == "mapping-incoming":
-                if inc_name in content_fields or inc_name.lower() in built_in_fields:
-                    return True
-            # outgoing mapper
-            if self.data.get('type', {}) == "mapping-outgoing":
-                # for inc timer type: "field.StartDate, and for using filters: "simple": "".
-                if simple := inc_info.get('simple'):
-                    if '.' in simple:
-                        simple = simple.split('.')[0]
-                    if simple in content_fields or simple in built_in_fields:
-                        return True
-            return False
-
-        return _extract_content_fields
-
-    def remove_inexistent_fields(self):
-        """
-        Remove in-existent fields from a mapper.
-        """
-        content_fields = get_all_incident_and_indicator_fields_from_id_set(self.id_set_file, 'mapper')
-        built_in_fields = [field.lower() for field in BUILT_IN_FIELDS] + LAYOUT_AND_MAPPER_BUILT_IN_FIELDS
+        content_fields = get_all_incident_and_indicator_fields_from_id_set(self.id_set_file, 'mapper') + [
+            field.lower() for field in BUILT_IN_FIELDS
+        ] + LAYOUT_AND_MAPPER_BUILT_IN_FIELDS
 
         mapper = self.data.get('mapping', {})
+        mapping_type = self.data.get('type', {})
+
         for mapping_name in mapper.values():
-            mapping_name['internalMapping'] = dict(
-                filter(
-                    self.extract_content_fields(content_fields=content_fields, built_in_fields=built_in_fields),
-                    mapping_name.get('internalMapping', {}).items()
+            internal_mapping_fields = mapping_name.get('internalMapping', {})
+            mapping_name['internalMapping'] = {
+                inc_name: inc_info for inc_name, inc_info in internal_mapping_fields.items()
+                if inc_name not in get_invalid_incident_fields_from_mapper(
+                    mapper_incident_fields=internal_mapping_fields,
+                    mapping_type=mapping_type,
+                    content_fields=content_fields,
                 )
-            )
+            }